product_service/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2-binary==2.9.6
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
redis==4.5.4
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
python-dotenv
django-redis==5.2.0
django-cors-headers==4.6.0